Revert "feat: เพิ่ม component test และ e2e test"
