--- backend-demo-test/.env
@@ -1,2 +1,0 @@
-MONGO_URL= mongodb+srv://user-123:pass123@mini-merm-tut.or0e1.mongodb.net/BestBeds?retryWrites=true&w=majority
-PASS_SEC = bestbeds
