+++ backend-demo-test/.gitignore
@@ -1,4 +1,4 @@
-
+.env
 .nyc_output
 build
 node_modules
