+++ backend-demo-test/src/index.js
@@ -16,7 +16,7 @@ app.use(cors());
 
 mongoose
     .connect(process.env.MONGO_URL)
-    .then(() => console.log(""))
+    .then(() => console.log("DB Connection Successfull!"))
     .catch((err) => {
         console.log(err);
     });
@@ -26,7 +26,7 @@ app.use(express.json());
 app.use(bedsRoute);
 app.use(bedsdealingRoute);
 app.use(authRoute);
-app.use('/users', userRoute);
+app.use(userRoute);
 
 app.listen(PORT, () => {
     console.log("Backend server is running! on port 8888")
